Adding a test framework and implementation of AuthenticationManager.
Adding an implementation of AuthenticationManager to support some
Pac4j-specific authentication mechanisms. Specifically, erasing
of credentials is not possible currently since Pac4j credentials
do not implement Spring Security's CredentialsContainer. Having
an implementation of AuthenticationManager that knows how to deal
with ClientAuthenticationTokens keeps credential erasing behavior
in the correct abstraction while not requiring it. It also gives
us an extension point for more Pac4j-specific functionality at
this level in the future.

* Added the implementation of AuthenticationManager

* Added Spock framework for unit testing